+++ LeetModel-backend/src/main/java/com/senior/leetmodelbackend/pojo/entity/OssFile.java
@@ -7,13 +7,13 @@ import java.time.LocalDateTime;
 @Data
 public class OssFile {
 
-    private Long fileId;
+    private Integer fileId;
     private String fileName;
     private String fileUrl;
     private String fileSuffix;
     private String contentType;
     private Long fileSize;
-    private Long uploaderId;
+    private Integer uploaderId;
     private Boolean isDeleted;
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
