+++ LeetModel-backend/src/main/java/com/senior/leetmodelbackend/pojo/entity/User.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @Data
 public class User {
 
-    private Long userId;
+    private Integer userId;
     private String username;
     private String email;
 
@@ -21,7 +21,7 @@ public class User {
     @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
     private String phone;
 
-    private Long avatarFileId;
+    private Integer avatarFileId;
     private String status;
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
